Add MagmaSend trait for Telegram message sending
A new trait called MagmaSend was added into the Magma4telegram package. This trait allows for the sending of Telegram messages via an API call. In addition, the MagmaKernel, Magma, and MagmaCommand files version numbers have been updated. Some minor modifications are made in the example modules to accommodate the new trait usage.