Refactor constructor in Magma class
The constructor of the Magma class has been refactored to get the command and chatId from the input package. Consequently, the extraction of these values has been removed from the api.php, simplifying the creation of a new Magma instance. Exception handling is also included within the constructor to ensure valid commands and chatIds.